requirements.txt
asgiref==3.8.1
Django==4.0.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-js-asset==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
mysql-connector-python==9.0.0
psycopg2-binary
django-filter==23.5
whitenoise==6.7.0
python-dotenv==1.0.1